+++ lib/number.dart
@@ -1,0 +1,5 @@
+/// Support for fractional and complex arithmetic.
+library more.number;
+
+export 'package:more/src/number/complex.dart' show Complex;
+export 'package:more/src/number/fraction.dart' show Fraction;